+++ src/components/index.ts
@@ -1,6 +1,5 @@
 export * from "./StreamMessage";
-export * from "./ToolWidgets"; 
-export * from "./NFOCredits"; 
+export * from "./ToolWidgets";
 export * from "./UsageDashboard";
 export * from "./WebviewPreview";
 export * from "./ImagePreview";